fix(influxdb_metrics sink): Fix formatting of timestamps in tests
